+++ command/agent/consul/sync_test.go
